Allow multiple modifiers
Prior to this change, only the last modifier got added to the sum of
rolls.

To fix this, the modifiers now get stored in an array and added to the
final sum one by one. For the message, it joins the the modifiers and
replaces all instances of "+ -" with "- " afterwards so it does not look
weird for negative modifiers