Add Request Queue
This plugin offers a simple queue for requests (or commissions) where
mods and admins can add or remove entries as they like. All requests can
be listed out. Out of all the requests, a random request can also be
chosen automatically.

The plugin listens to the special event `requestq_addRequestToQ` on the
`api.Events` EventEmitter and takes a single argument, namely the
request to be added as string